+++ RELEASES.md
@@ -1,5 +1,25 @@
 # Confidential Match Source Code Releases
 
+## September 29, 2025
+
+* Version: v1.267.0
+* Lookup Service
+    * Image Digest: sha256:f2f9d08ef228d1cdc1f9cdd0893d8cfdd24ac61e518f11d13c517c7b7e1efba5
+    * [Signature](https://us-docker.pkg.dev/admcloud-cfm-public/docker-repo-signatures/lookup_server_gcp_signature@sha256:c8d2c9c311959e7d0b853d3062852b84104c3f1b64e1d1fceb41e95fd0e6584d)
+* Match Request Processor
+    * Image Digest: sha256:e818ce3a7d449b18e4847fee1fd5a90a8ff44a3ed08375bbe40726e17167ec0a
+    * [Signature](https://us-docker.pkg.dev/admcloud-cfm-public/docker-repo-signatures/mrp_app_gcp_signature@sha256:13b263bf62da66eefeb1fe3f2caac54f83b69bb5e2210d82b9c0636eec942821)
+
+## September 24, 2025
+
+* Version: v1.264.0
+* Lookup Service
+    * Image Digest: sha256:f2f9d08ef228d1cdc1f9cdd0893d8cfdd24ac61e518f11d13c517c7b7e1efba5
+    * [Signature](https://us-docker.pkg.dev/admcloud-cfm-public/docker-repo-signatures/lookup_server_gcp_signature@sha256:c8d2c9c311959e7d0b853d3062852b84104c3f1b64e1d1fceb41e95fd0e6584d)
+* Match Request Processor
+    * Image Digest: sha256:a82da72227ed9c10e9441b340a5f90295b02209436ee792c748c0a092cbe27ec
+    * [Signature](https://us-docker.pkg.dev/admcloud-cfm-public/docker-repo-signatures/mrp_app_gcp_signature@sha256:80df34276a7404f5f2facfae3721146734437b75c262a7ed96d784527bdaeda1)
+
 ## September 18, 2025
 
 * Version: v1.260.0